Tests: * Migrate Kotlin List Method Tests to Lists-ktx Module * Convert DoubleListLargeSumTest.java to DoubleListLargeSumTest.kt * Move ObjectListTest from Objects package to Lists package
