+++ lists-ktx/src/test/kotlin/mathtools/lists/bytes/ByteListFindOutOfRangeTest.kt
