+++ lists-ktx/src/test/kotlin/mathtools/lists/bytes/ByteListLargeSumTest.kt
@@ -1,34 +1,36 @@
-package mathtools.lists.ints
+package mathtools.lists.bytes
 
-import mathtools.lists.IntList.largeSum
-import mathtools.lists.NumberListConversion.toInt
+import mathtools.lists.ByteList.largeSum
+import mathtools.lists.NumberListConversion.toByte
 import mathtools.lists.testdata.ListDataSource.uniform101
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import java.math.BigInteger
 
-/** Testing the IntList LargeSum function
+/** Testing [ByteList] LargeSum function
  * @author DK96-OS : 2022 */
-class IntListLargeSumTest {
+class ByteListLargeSumTest {
 
-	private val limit = Integer.MAX_VALUE / 2
+	private val limit: Byte = (Byte.MAX_VALUE / 2).toByte()
 	private val bigLimit = BigInteger.valueOf(limit.toLong())
 
 	@Test
 	fun testU101() {
-		val data = toInt(uniform101)
+		val data = toByte(uniform101)
 		assertEquals(
-			BigInteger.valueOf(3030), largeSum(data)
+			BigInteger.valueOf(3030),
+			largeSum(data)
 		)
 	}
 
 	@Test
 	fun testMaxIntValues() {
-		val data = mutableListOf<Int>()
-		for (i in 0 until 5000)
-			data.add(Int.MAX_VALUE)
+		val data = mutableListOf<Byte>()
+		for (i in 0 until 50)
+			data.add(Byte.MAX_VALUE)
 		assertEquals(
-			BigInteger.valueOf(5000L * Int.MAX_VALUE), largeSum(data)
+			BigInteger.valueOf(50L * Byte.MAX_VALUE),
+			largeSum(data)
 		)
 	}
 
@@ -41,14 +43,14 @@ class IntListLargeSumTest {
 			bigLimit, largeSum(listOf(limit)))
 		//
 		assertEquals(
-			bigLimit.multiply(BigInteger.TWO),
+			bigLimit.multiply(BigInteger.valueOf(2)),
 			largeSum(listOf(limit, limit))
 		)
 	}
 
 	@Test
 	fun testNegativeValues() {
-		val value = -(limit + 1) / 2
+		val value: Byte = (-(limit + 1) / 2).toByte()
 		val array = Array(8) { value }
 		//
 		assertEquals(
