+++ lists-ktx/src/test/kotlin/mathtools/lists/bytes/ByteListRemoveByIndicesTest.kt
