+++ lists-ktx/src/test/kotlin/mathtools/lists/doubles/DoubleListFindGreaterThanTest.kt
