+++ lists-ktx/src/test/kotlin/mathtools/lists/doubles/DoubleListFindLessThanTest.kt
