+++ lists-ktx/src/test/kotlin/mathtools/lists/doubles/DoubleListFindOutOfBoundsTest.kt
