+++ lists-ktx/src/test/kotlin/mathtools/lists/doubles/DoubleListRemoveByIndicesTest.kt
