+++ lists-ktx/src/test/kotlin/mathtools/lists/ints/IntListFindOutOfRangeTest.kt
