+++ lists-ktx/src/test/kotlin/mathtools/lists/longs/LongListFindOutOfRangeTest.kt
